+++ numba/datamodel/__init__.py
@@ -1,0 +1,4 @@
+from .manager import DataModelManager
+from .funcinfo import FunctionInfo
+from .registry import *
+from .models import *
